refactor(application): relocate vulnerability_builder and resolution_guide_builder tests to their modules
Move 9 vulnerability_builder unit tests (map_severity, build_vulnerability_view,
build_vulnerabilities) and 3 resolution_guide_builder unit tests
(build_resolution_guide) from mod.rs to their respective submodule files.

This completes the final subtask of #530: mod.rs now contains exactly the 8
integration tests that exercise SbomReadModelBuilder end-to-end, with all
unit tests placed alongside the functions they test.

Closes #533
