+++ src/application/read_models/sbom_read_model_builder/resolution_guide_builder.rs
@@ -43,6 +43,8 @@ fn build_resolution_entry_view(entry: &ResolutionEntry) -> ResolutionEntryView {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::application::read_models::sbom_read_model_builder::test_helpers as th;
+    use crate::application::read_models::vulnerability_view::SeverityView;
     use crate::sbom_generation::domain::resolution_guide::{IntroducedBy, ResolutionEntry};
     use crate::sbom_generation::domain::vulnerability::Severity;
 
@@ -96,5 +98,65 @@ mod tests {
         assert_eq!(guide.entries.len(), 2);
         assert_eq!(guide.entries[0].dependency_chains.len(), 1);
         assert!(guide.entries[1].dependency_chains.is_empty());
+    }
+
+    #[test]
+    fn test_build_resolution_guide_converts_entries() {
+        let entries = vec![th::resolution_entry(
+            "urllib3",
+            "1.26.5",
+            Some("1.26.18"),
+            Severity::High,
+            "CVE-2023-43804",
+            &[("requests", "2.28.0")],
+        )];
+
+        let guide = build_resolution_guide(&entries);
+
+        assert_eq!(guide.entries.len(), 1);
+        assert_eq!(guide.entries[0].vulnerable_package, "urllib3");
+        assert_eq!(guide.entries[0].current_version, "1.26.5");
+        assert_eq!(guide.entries[0].fixed_version, Some("1.26.18".to_string()));
+        assert_eq!(guide.entries[0].severity, SeverityView::High);
+        assert_eq!(guide.entries[0].vulnerability_id, "CVE-2023-43804");
+        assert_eq!(guide.entries[0].introduced_by.len(), 1);
+        assert_eq!(guide.entries[0].introduced_by[0].package_name, "requests");
+        assert_eq!(guide.entries[0].introduced_by[0].version, "2.28.0");
+    }
+
+    #[test]
+    fn test_build_resolution_guide_without_fixed_version() {
+        let entries = vec![th::resolution_entry(
+            "vulnerable-pkg",
+            "0.1.0",
+            None,
+            Severity::Critical,
+            "CVE-2024-0001",
+            &[("parent-pkg", "1.0.0")],
+        )];
+
+        let guide = build_resolution_guide(&entries);
+
+        assert_eq!(guide.entries.len(), 1);
+        assert_eq!(guide.entries[0].fixed_version, None);
+        assert_eq!(guide.entries[0].severity, SeverityView::Critical);
+    }
+
+    #[test]
+    fn test_build_resolution_guide_multiple_introduced_by() {
+        let entries = vec![th::resolution_entry(
+            "urllib3",
+            "1.26.5",
+            Some("1.26.18"),
+            Severity::High,
+            "CVE-2023-43804",
+            &[("httpx", "0.23.0"), ("requests", "2.28.0")],
+        )];
+
+        let guide = build_resolution_guide(&entries);
+
+        assert_eq!(guide.entries[0].introduced_by.len(), 2);
+        assert_eq!(guide.entries[0].introduced_by[0].package_name, "httpx");
+        assert_eq!(guide.entries[0].introduced_by[1].package_name, "requests");
     }
 }
